+++ changes.md
@@ -1,3 +1,6 @@
-#### 0.0.1
+#### 0.0.2
+* FIX: improved compatibility with Stateless mode.
+* FIX: get the correct list of files for Compatibility Files Sync.
 
-- Initial public release.
+#### 0.0.1
+* Initial public release.